done filter product

+++ app/Http/Controllers/ProductController.php
@@ -3,13 +3,13 @@
 namespace App\Http\Controllers;
 
 use App\Models\Branch;
-use App\Models\Cate_Product;
 use App\Models\Category;
+use App\Models\Cate_Product;
 use App\Models\Product;
 use App\Models\Tag;
 use App\Models\Tag_Product;
 use Illuminate\Http\Request;
-
+use Illuminate\Support\Facades\Route;
 class ProductController extends Controller
 {
     /**
@@ -17,11 +17,61 @@ class ProductController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
-    {
-        $models = Product::all();
+    protected $categories;
+    protected $branches;
+    protected $tags;
 
-        return view('admin.pages.products.index', compact('models'));
+    public function __construct()
+    {
+        $this->categories = Category::select(['id','slug', 'name'])->get();
+        $this->branches = Branch::select(['id', 'slug', 'name'])->get();
+        $this->tags = Tag::select(['id', 'slug', 'name'])->get();
+    }
+    public function index(Request $request)
+    {
+        $models = Product::where(function ($query) use ($request) {
+            if ($request->category) {
+                $cate = Category::where('slug', $request->category)->first();
+                $items = $cate->cate_products;
+                $pro_id = [];
+                foreach ($items as $i) {
+                    $pro_id[] = $i->product_id;
+                }
+                return $query->from('products')->whereIn('id', $pro_id);
+            }
+        })->where(function($query) use($request){
+            if ($request->tag) {
+                $tag = Tag::where('slug', $request->tag)->first();
+                $items = $tag->tag_products;
+                $pro_id = [];
+                foreach ($items as $i) {
+                    $pro_id[] = $i->product_id;
+                }
+                return $query->from('products')->whereIn('id', $pro_id);
+            }
+        })->where(function($query) use($request){
+            return $request->branch?$query->from('products')->where('branch_id', $request->branch) : '';
+        })->where(function($query) use($request){
+            return $request->keyword?$query->from('products')->where('name', 'like', "%$request->keyword%"):'';
+        })
+        ->orderBy('id', 'desc')
+            ->paginate(5);
+
+        $categories = $this->categories;
+        $branches = $this->branches;
+        $tags = $this->tags;
+        $request->session()->flash('check', ['cate'=> $request->category, 'tag'=>$request->tag, 'branch'=>$request->branch]);
+        // pagination
+        $param = [
+            'tag'=>$request->tag,
+            'category'=>$request->category,
+            'branch'=>$request->branch,
+            'keyword'=>$request->keyword,
+        ];
+        $totalPage = $models->lastPage();
+        $currentPage = $models->currentPage();
+        return view('admin.pages.products.index', 
+        compact('models', 'categories', 'branches', 'tags', 'totalPage', 'currentPage', 'param'));
     }
 
     /**
@@ -30,10 +80,10 @@ class ProductController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function create()
-    {   
-        $categories = Category::all();
-        $branches = Branch::all();
-        $tags = Tag::all();
+    {
+        $categories = $this->categories;
+        $branches = $this->branches;
+        $tags = $this->tags;
         return view('admin.pages.products.create', compact('categories', 'branches', 'tags'));
     }
 
@@ -44,18 +94,18 @@ class ProductController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
-    {   
+    {
         $model = new Product();
         $request->validate([
-            'name'=>'required|unique:products',
-            'slug'=>'required|unique:products',
-            'image'=>'required|mimes:jpg,bmp,png',
-            'price'=>'required',
-            'competitive_price'=>'required',
-            'short_desc'=>'required'
+            'name' => 'required|unique:products',
+            'slug' => 'required|unique:products',
+            'image' => 'required|mimes:jpg,bmp,png',
+            'price' => 'required',
+            'competitive_price' => 'required',
+            'short_desc' => 'required',
         ]);
-        
-        if($request->hasFile('image')){
+
+        if ($request->hasFile('image')) {
             $destinationPath = 'public/images/products';
             $image = $request->file('image');
             $image_name = $image->getClientOriginalName();
@@ -67,14 +117,18 @@ class ProductController extends Controller
         $model->fill($_POST);
         $model->image = $request->image;
         $model->save();
-        for($i = 0; $i < count($_POST['cate_id']); $i++){
-            $arr1[] = ['cate_id'=>$_POST['cate_id'][$i], 'product_id'=>$model->id];
+        if (isset($_POST['cate_id'])) {
+            for ($i = 0; $i < count($_POST['cate_id']); $i++) {
+                $arr1[] = ['cate_id' => $_POST['cate_id'][$i], 'product_id' => $model->id];
+            }
+            Cate_Product::insert($arr1);
         }
-        for($i = 0; $i < count($_POST['tag_id']); $i++){
-            $arr2[] = ['tag_id'=>$_POST['tag_id'][$i], 'product_id'=>$model->id];
+        if (isset($_POST['cate_id'])) {
+            for ($i = 0; $i < count($_POST['tag_id']); $i++) {
+                $arr2[] = ['tag_id' => $_POST['tag_id'][$i], 'product_id' => $model->id];
+            }
+            Tag_Product::insert($arr2);
         }
-        Cate_Product::insert($arr1);
-        Tag_Product::insert($arr2);
         $request->session()->flash('message', 'Thêm sản phẩm thành công');
         return redirect('/admin/products');
     }
@@ -102,12 +156,12 @@ class ProductController extends Controller
         $product_ids = $model->product_cate_pros;
         $tag_ids = $model->product_tags;
 
-        if($model){
-            $categories = Category::all();
-            $branches = Branch::all();
-            $tags = Tag::all();
-            return view('admin.pages.products.edit', compact('model', 'categories', 'branches', 'tags','product_ids', 'tag_ids'));
-        }else{
+        if ($model) {
+            $categories = $this->categories;
+            $branches = $this->branches;
+            $tags = $this->tags;
+            return view('admin.pages.products.edit', compact('model', 'categories', 'branches', 'tags', 'product_ids', 'tag_ids'));
+        } else {
             return redirect('/admin/products');
         }
     }
@@ -120,49 +174,55 @@ class ProductController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request)
-    {
+    {   
+
         $request->validate([
-            'name'=>'required',
-            'slug'=>'required',
-            'image'=>'mimes:jpg,bmp,png',
-            'price'=>'required',
-            'competitive_price'=>'required',
-            'short_desc'=>'required'
+            'name' => 'required',
+            'slug' => 'required',
+            'image' => 'mimes:jpg,bmp,png',
+            'price' => 'required',
+            'competitive_price' => 'required',
+            'short_desc' => 'required',
         ]);
         $image = '';
         $model = Product::find($request->product_id);
-        if($request->hasFile('image')){
+        if ($request->hasFile('image')) {
             $destinationPath = 'public/images/products';
             $image = $request->file('image');
             $image_name = $image->getClientOriginalName();
             $request->file('image')->storeAs($destinationPath, $image_name);
             $image = $image_name;
-            if(file_exists('storage/images/products/'.$model->image)){
-                unlink('storage/images/products/'.$model->image);
+            if (file_exists('storage/images/products/' . $model->image)) {
+                unlink('storage/images/products/' . $model->image);
             }
         } else {
             $image = $request->input('image2');
         }
-        if($model){
+        if ($model) {
             $arr1 = [];
             $arr2 = [];
             $model->fill($_POST);
             $model->image = $image;
             $model->save();
-            for($i = 0; $i < count($_POST['cate_id']); $i++){
-                $arr1[] = ['cate_id'=>$_POST['cate_id'][$i], 'product_id'=>$model->id];
+            if (isset($_POST['cate_id'])) {
+                for ($i = 0; $i < count($_POST['cate_id']); $i++) {
+                    $arr1[] = ['cate_id' => $_POST['cate_id'][$i], 'product_id' => $model->id];
+                }
+                Cate_Product::where('product_id', $model->id)->delete();
+                Cate_Product::insert($arr1);
             }
-            for($i = 0; $i < count($_POST['tag_id']); $i++){
-                $arr2[] = ['tag_id'=>$_POST['tag_id'][$i], 'product_id'=>$model->id];
-            }
-            Cate_Product::where('product_id', $model->id)->delete();
-            Cate_Product::insert($arr1);
 
-            Tag_Product::where('product_id', $model->id)->delete();
-            Tag_Product::insert($arr2);
+            if (isset($_POST['tag_id'])) {
+                for ($i = 0; $i < count($_POST['tag_id']); $i++) {
+                    $arr2[] = ['tag_id' => $_POST['tag_id'][$i], 'product_id' => $model->id];
+                }
+                Tag_Product::where('product_id', $model->id)->delete();
+                Tag_Product::insert($arr2);
+            }
+
             $request->session()->flash('message', 'Cập nhật sản phẩm thành công');
             return redirect('/admin/products');
-        }else{
+        } else {
             return redirect('/admin/products');
         }
     }
@@ -176,14 +236,14 @@ class ProductController extends Controller
     public function destroy($id, Request $request)
     {
         $model = Product::find($id);
-        if($model){
-            if(file_exists('storage/images/products/'.$model->image)){
-                unlink('storage/images/products/'.$model->image);
+        if ($model) {
+            if (file_exists('storage/images/products/' . $model->image)) {
+                unlink('storage/images/products/' . $model->image);
             }
             $model->delete();
             $request->session()->flash('message', 'Xóa sản phẩm thành công');
             return redirect('/admin/products');
-        }else{
+        } else {
             return redirect('/admin/products');
         }
     }
